Transcription-Diarization-Bot/requirements.txt
fastapi
uvicorn
openai-whisper
pyannote.audio
torch
torchaudio
numpy
pydub
gemini-ai
ffmpeg
requests
python-dotenv
